+++ src/OpenTelemetryBundle.php
@@ -11,7 +11,7 @@ use Traceway\OpenTelemetryBundle\DependencyInjection\Compiler\HttpClientTracingP
 
 final class OpenTelemetryBundle extends Bundle
 {
-    public const VERSION = '1.3.0';
+    public const VERSION = '1.3.1';
 
     public function getPath(): string
     {